Avoid warnings from bare trait objects
Since we support Rust 1.18 but modern Rust versions have a warning
about bare trait objects, we go ahead and allow them in the code of
`src/lib.rs` & `snafu-derive/src/lib.rs`. We also have to allow
unknown lints because Rust 1.18 doesn't know about the lint we are
allowing.

We can use `dyn` in the documentation and tests because those are
only executed in modern Rust versions.