Implement support for no_std
This brings no_std support to SNAFU. It now has a `std` feature which
is activated by default. To make no_std support as frictionless as
possible, SNAFU now reexports the `std::error::Error` trait as
`snafu::Error` when the feature is activated and defines its own API
compatible trait instead when it's disabled.

In order to not commit to having a publicly visible copy of the
`Error` trait in SNAFU, hiding it from the documentation allows for
SNAFU to still support all of its features, but the user never gets to
use the trait. This way SNAFU can switch out the trait at any point
with `core::error::Error` if that becomes a thing. Alternatively the
trait here can also be made visible at some point. Both of these
possibilities are not breaking, so the conservative approach with
`doc(hidden)` allows for the smoothest experience going forward.

Resolves #85