Fix parsing of source/backtrace with no parens
If a plain `source` or `backtrace` attribute were specified in a non-final
position in an attribute list, we would fail to parse them because we assumed
`input` would be empty if we were at the end of an attribute.  However, `input`
is the parse stream for the whole attribute list, not just the attribute; it
could contain a comma and further attributes.

This change checks more carefully whether we're at the end of an attribute by
looking for empty input or a comma that signifies another attribute is coming.
Error messages are also improved, because it uses `Lookahead1` to let syn track
the tokens we want to allow.

Fixes #146