Remove the rust_1_30 feature and test for 1.31 compatibility
The new minimum supported Rust version will be 1.31.  This removes the
rust_1_30 feature, which was used to opt into better support for
higher-than-minimum versions of Rust; it's no longer needed because there we
have no features that require beyond Rust 1.31.  Code that was conditional on
rust_1_30 is now required.

Compatibility tests for 1.30 were replaced with 1.31.

Background: https://github.com/shepmaster/snafu/issues/154