Restructure parsing of `Snafu` attributes and fields
This largely started out as moving code out of `lib.rs` and into
smaller, more focused files. Along the way, a number of
changes (sometimes even improvements) were made.

- Instead of tracking a vector of `syn::Errors`, use its built-in
  ability to `combine` errors.

- Remove some intermediate representations of the
  attributes. Instead, we use the parsed attributes themselves when
  possible.

- Reduce some allocations for error messages.

- Improved the spans of some error messages.

- Made some error messages more specific. For example, we now mention `source(from)` instead
  of `source`.

- Improved the general wording of some error messages.

- Added some abstractions for attribute metadata. This helps avoid
  copy-pasting attribute names around.

This rewriting does split apart some errors that were previously
reported in the same phase. In practice, this is unlikely to matter as
most of the time people aren't making every possible error at the same
time. Where possible, I tried to keep the number of phases as low as
possible to allow more errors to be reported at once.